tier/dht : Default value for demote-freq, max files and mb
Default value for tier-demote-frequency is 3600 sec to avoid
frequent demotions.

Default value for tier-max-mb is 4000 mb

Default value for tier-max-files is 10000 files

BUG: 1300412
Reviewed-on: http://review.gluster.org/13270
Smoke: Gluster Build System <jenkins@build.gluster.com>
Tested-by: Dan Lambright <dlambrig@redhat.com>
NetBSD-regression: NetBSD Build System <jenkins@build.gluster.org>
CentOS-regression: Gluster Build System <jenkins@build.gluster.com>

+++ xlators/cluster/dht/src/tier.c
@@ -1841,7 +1841,7 @@ tier_init (xlator_t *this)
                 freq = DEFAULT_TIER_MAX_MIGRATE_MB;
         }
 
-        defrag->tier_conf.max_migrate_bytes = freq * 1024 * 1024;
+        defrag->tier_conf.max_migrate_bytes = (uint64_t) freq * 1024 * 1024;
 
         ret = dict_get_int32 (this->options,
                               "tier-max-files", &freq);
@@ -2018,7 +2018,8 @@ tier_reconfigure (xlator_t *this, dict_t *options)
                 GF_OPTION_RECONF ("tier-max-mb",
                                   migrate_mb, options,
                                   int32, out);
-                defrag->tier_conf.max_migrate_bytes = migrate_mb*1024*1024;
+                defrag->tier_conf.max_migrate_bytes = (uint64_t) migrate_mb *
+                                                        1024 * 1024;
 
                 GF_OPTION_RECONF ("tier-max-files",
                                   defrag->tier_conf.max_migrate_files, options,
